Make a schema package
