+++ recipe/schemas/config_schemas.py
@@ -2,78 +2,19 @@
 Registers recipe schemas
 """
 
-import inspect
 import logging
 import re
 
 from copy import copy
-from sqlalchemy import distinct, func
 from sureberus import schema as S
-from sureberus import errors as E
-
 from recipe.compat import basestring
-from lark import Lark, Transformer, v_args
+from .utils import sqlalchemy_datatypes, coerce_format, aggregations, SCALAR_TYPES, pop_version
 
 logging.captureWarnings(True)
 
-SCALAR_TYPES = [S.Integer(), S.String(), S.Float(), S.Boolean()]
-
-
-def _make_sqlalchemy_datatype_lookup():
-    """ Build a dictionary of the allowed sqlalchemy casts """
-    from sqlalchemy.sql import sqltypes
-
-    d = {}
-    for name in dir(sqltypes):
-        sqltype = getattr(sqltypes, name)
-        if name.lower() not in d and name[0] != "_" and name != "NULLTYPE":
-            if inspect.isclass(sqltype) and issubclass(sqltype, sqltypes.TypeEngine):
-                d[name.lower()] = sqltype
-    return d
-
-
-sqlalchemy_datatypes = _make_sqlalchemy_datatype_lookup()
-
-format_lookup = {
-    "comma": ",.0f",
-    "dollar": "$,.0f",
-    "percent": ".0%",
-    "comma1": ",.1f",
-    "dollar1": "$,.1f",
-    "percent1": ".1%",
-    "comma2": ",.2f",
-    "dollar2": "$,.2f",
-    "percent2": ".2%",
-}
-
 default_aggregation = "sum"
 no_aggregation = "none"
-aggregations = {
-    "sum": func.sum,
-    "min": func.min,
-    "max": func.max,
-    "avg": func.avg,
-    "count": func.count,
-    "count_distinct": lambda fld: func.count(distinct(fld)),
-    "month": lambda fld: func.date_trunc("month", fld),
-    "week": lambda fld: func.date_trunc("week", fld),
-    "year": lambda fld: func.date_trunc("year", fld),
-    "quarter": lambda fld: func.date_trunc("quarter", fld),
-    "age": lambda fld: func.date_part("year", func.age(fld)),
-    "none": lambda fld: fld,
-    None: lambda fld: fld,
-    # Percentile aggregations do not work in all engines
-    "median": func.median,
-    "percentile1": lambda fld: func.percentile_cont(0.01).within_group(fld),
-    "percentile5": lambda fld: func.percentile_cont(0.05).within_group(fld),
-    "percentile10": lambda fld: func.percentile_cont(0.10).within_group(fld),
-    "percentile25": lambda fld: func.percentile_cont(0.25).within_group(fld),
-    "percentile50": lambda fld: func.percentile_cont(0.50).within_group(fld),
-    "percentile75": lambda fld: func.percentile_cont(0.75).within_group(fld),
-    "percentile90": lambda fld: func.percentile_cont(0.90).within_group(fld),
-    "percentile95": lambda fld: func.percentile_cont(0.95).within_group(fld),
-    "percentile99": lambda fld: func.percentile_cont(0.99).within_group(fld),
-}
+
 
 aggr_keys = "|".join(k for k in aggregations.keys() if isinstance(k, basestring))
 # Match patterns like sum(a)
@@ -472,10 +413,6 @@ def _replace_references(shelf):
     return shelf
 
 
-def coerce_format(v):
-    return format_lookup.get(v, v)
-
-
 condition_schema = _full_condition_schema(aggr=False)
 
 quickselect_schema = S.List(
@@ -535,306 +472,10 @@ ingredient_schema = S.Dict(
 )
 
 
-def pop_version(d):
-    d.pop("_version", None)
-    return d
-
-
-shelf_schema_v1 = S.Dict(
+shelf_schema = S.Dict(
     valueschema=ingredient_schema,
     keyschema=S.String(),
     allow_unknown=True,
     coerce=pop_version,
     coerce_post=_replace_references,
-)
-
-
-############################
-# PARSED
-# Shelf config _version="2" supports parsed fields.
-############################
-
-field_grammar = """
-    ?start:  expr | bool_expr | partial_relation_expr
-    ?expr: agex | sum                          -> expr
-    ?agex: aggr "(" sum ")"
-        | /count/i "(" STAR ")"                -> agex
-    ?aggr:  /(sum|min|max|avg|count)/i         -> aggregate
-    ?sum: product
-        | sum "+" product                      -> add
-        | sum "-" product                      -> sub
-    ?product: atom
-           | product "*" atom                  -> mul
-           | product "/" atom                  -> div
-    ?atom: agex
-           | const
-           | column
-           | case
-           | "(" sum ")"
-    ?column:  NAME                             -> column
-    ?const: NUMBER                             -> number
-            | ESCAPED_STRING                   -> literal
-            | /true/i                          -> true
-            | /false/i                         -> false
-            | /null/i                          -> null
-
-    // Pairs of boolean expressions and expressions
-    // forming case when {BOOL_EXPR} then {EXPR}
-    // an optional final expression is the else.T
-    ?case: "if" "(" (bool_expr "," expr)+ ("," expr)? ")"
-
-    // boolean expressions
-    ?bool_expr: bool_term ["OR" bool_term]
-    ?bool_term: bool_factor ["AND" bool_factor]
-    ?bool_factor: column
-                  | "NOT" bool_factor          -> not_bool_factor
-                  | "(" bool_expr ")"
-                  | relation_expr
-                  | vector_relation_expr
-    ?partial_relation_expr: comparator atom
-                | vector_comparator array
-                | BETWEEN pair_array
-    ?relation_expr:        atom comparator atom
-    ?vector_relation_expr: atom vector_comparator array
-                         | atom BETWEEN pair_array
-    ?pair_array:           "(" const "," const ")"      -> array
-    ?array:                "(" [const ("," const)*] ")"
-    ?comparator: EQ | NE | LT | LTE | GT | GTE
-    ?vector_comparator: IN | NOTIN
-    EQ: "="
-    NE: "!="
-    LT: "<"
-    LTE: "<="
-    GT: ">"
-    GTE: ">="
-    IN: /IN/i
-    NOTIN: /NOT/i /IN/i
-    BETWEEN: /BETWEEN/i
-    STAR: "*"
-    COMMENT: /#.*/
-
-    %import common.CNAME                       -> NAME
-    %import common.SIGNED_NUMBER               -> NUMBER
-    %import common.ESCAPED_STRING
-    %import common.WS_INLINE
-    %ignore COMMENT
-    %ignore WS_INLINE
-"""
-
-field_parser = Lark(field_grammar, parser="earley")
-
-
-def parsed_move_fields(value):
-    """ Move any fields that look like "{role}_field" into the extra_fields
-    list. These will be processed as fields. Rename them as {role}_expression.
-    """
-    if isinstance(value, dict):
-        keys_to_move = [k for k in value.keys() if k.endswith("_field")]
-        if keys_to_move:
-            value["extra_fields"] = []
-            for k in keys_to_move:
-                value["extra_fields"].append(
-                    {"name": k[:-6] + "_expression", "field": value.pop(k)}
-                )
-
-    return value
-
-
-def add_version(v):
-    # Add version to a parsed ingredient
-    v["_version"] = "2"
-    return v
-
-
-import attr
-
-
-@attr.s
-class ParseValidator:
-    #: Message to display on failure
-    msg = attr.ib()
-    #: Must begin with one of these tokens
-    required_head_tokens = attr.ib(default=[])
-    #: Must not contain any of these tokens anywhere
-    forbidden_tokens = attr.ib(default=[])
-    #: Must contain at least one of these tokens
-    required_tokens = attr.ib(default=[])
-    other = attr.ib(default=None)
-
-    def __and__(self, other):
-        self.other = other
-        return self
-
-    def __call__(self, f, v, e):
-        """Check parsing"""
-        try:
-            tree = field_parser.parse(v)
-            failure_message = str(v) + " " + self.msg
-
-            if self.required_head_tokens:
-                if tree.data not in self.required_head_tokens:
-                    raise e(f, failure_message)
-            if self.forbidden_tokens:
-                for tok in self.forbidden_tokens:
-                    if list(tree.find_data(tok)):
-                        raise e(f, failure_message)
-            if self.required_tokens:
-                for tok in self.required_tokens:
-                    if not list(tree.find_data(tok)):
-                        raise e(f, failure_message)
-        except Exception as exc:
-            # A Lark error message raised when the value doesn't parse
-            raise exc
-
-        # Validate against the other
-        if self.other:
-            self.other(f, v, e)
-
-
-validate_parses_with_agex = ParseValidator(
-    msg="must contain an aggregate expression", required_tokens=["agex"]
-)
-validate_parses_without_agex = ParseValidator(
-    msg="must not contain an aggregate expression", forbidden_tokens=["agex"]
-)
-validate_partial_or_full_condition = ParseValidator(
-    msg="must be a condition or partial condition",
-    required_head_tokens=[
-        "partial_relation_expr",
-        "bool_expr",
-        "bool_term",
-        "bool_factor",
-        "relation_expr",
-    ],
-) & validate_parses_without_agex
-validate_condition = ParseValidator(
-    msg="must be a condition",
-    required_head_tokens=["bool_expr", "bool_term", "bool_factor", "relation_expr"],
-) & validate_parses_without_agex
-validate_agex_condition = ParseValidator(
-    msg="must be a condition",
-    required_head_tokens=["bool_expr", "bool_term", "bool_factor", "relation_expr"],
-) & validate_parses_with_agex
-
-
-parsed_condition_schema = S.String(required=True, validator=validate_condition)
-
-parsed_full_or_partial_condition_schema = S.String(
-    required=True, validator=validate_partial_or_full_condition
-)
-
-labeled_parsed_condition_schema = S.Dict(
-    schema={"condition": parsed_condition_schema, "label": S.String(required=True)}
-)
-
-parsed_quickselect_schema = S.List(
-    required=False, schema=S.Dict(schema=labeled_parsed_condition_schema)
-)
-
-parsed_agex_field_schema = S.String(required=True, validator=validate_parses_with_agex)
-
-parsed_noag_field_schema = S.String(
-    required=True, validator=validate_parses_without_agex
-)
-
-parsed_format_schema = S.String(coerce=coerce_format, required=False)
-
-parsed_metric_schema = S.Dict(
-    schema={
-        "field": parsed_agex_field_schema,
-        "format": parsed_format_schema,
-        "quickselects": parsed_quickselect_schema,
-    },
-    coerce_post=add_version,
-    allow_unknown=True,
-)
-
-parsed_dimension_schema = S.Dict(
-    allow_unknown=True,
-    coerce=parsed_move_fields,
-    coerce_post=add_version,
-    schema={
-        "field": parsed_noag_field_schema,
-        "extra_fields": S.List(
-            required=False,
-            schema=S.Dict(
-                schema={
-                    "field": parsed_noag_field_schema,
-                    "name": S.String(required=True),
-                }
-            ),
-        ),
-        "buckets": S.List(
-            required=False, schema=parsed_full_or_partial_condition_schema
-        ),
-        "buckets_default_label": {"anyof": SCALAR_TYPES, "required": False},
-        "format": parsed_format_schema,
-        "quickselects": parsed_quickselect_schema,
-    },
-)
-
-parsed_filter_schema = S.Dict(
-    allow_unknown=True,
-    coerce_post=add_version,
-    schema={"condition": parsed_condition_schema},
-)
-
-parsed_having_schema = S.Dict(
-    allow_unknown=True,
-    coerce_post=add_version,
-    schema={"condition": parsed_condition_schema},
-)
-
-parsed_ingredient_schema_choices = {
-    "Metric": parsed_metric_schema,
-    "Dimension": parsed_dimension_schema,
-    "Filter": parsed_filter_schema,
-    "Having": parsed_having_schema,
-}
-
-
-parsed_ingredient_schema = S.Dict(
-    choose_schema=S.when_key_is(
-        "kind",
-        {
-            "Metric": parsed_metric_schema,
-            "Dimension": parsed_dimension_schema,
-            "Filter": parsed_filter_schema,
-            "Having": parsed_having_schema,
-        },
-        default_choice="Metric",
-    ),
-    registry={},
-)
-
-parsed_shelf_schema = S.Dict(
-    valueschema=parsed_ingredient_schema,
-    keyschema=S.String(),
-    coerce=pop_version,
-    allow_unknown=True,
-)
-
-
-shelf_schema = S.Dict(
-    choose_schema=S.when_key_is(
-        "_version", {"1": shelf_schema_v1, "2": parsed_shelf_schema}, default_choice="1"
-    )
-)
-
-# This schema is used with sureberus
-recipe_schema = S.Dict(
-    schema={
-        "metrics": S.List(schema=S.String(), required=False),
-        "dimensions": S.List(schema=S.String(), required=False),
-        "filters": S.List(schema={"oneof": [S.String(), "condition"]}, required=False),
-        "order_by": S.List(schema=S.String(), required=False),
-    },
-    registry={
-        "aggregated_field": _field_schema(aggr=True, required=True),
-        "optional_aggregated_field": _field_schema(aggr=True, required=False),
-        "non_aggregated_field": _field_schema(aggr=False, required=True),
-        "condition": _full_condition_schema(aggr=False, label_required=False),
-        "labeled_condition": _full_condition_schema(aggr=False, label_required=True),
-        "having_condition": _full_condition_schema(aggr=True, label_required=False),
-    },
 )
