+++ tests/parsed_ingredients/census_complex.yaml
@@ -1,20 +1,20 @@
 state:
     kind: Dimension
-    parsed_field: state
+    field: state
     lookup:
         Vermont: "The Green Mountain State"
         Tennessee: "The Volunteer State"
     quickselects:
         - name: younger
-          parsed_condition: "age<40"
+          condition: "age<40"
         - name: vermontier
-          parsed_condition: "state='Vermont'"
+          condition: "='Vermont'"
 pop2000:
     kind: Metric
-    parsed_field: "if(age > 40, pop2000)"
+    field: "if(age > 40, pop2000)"
 pop2008:
     kind: Metric
-    parsed_field: pop2008
+    field: pop2008
 allthemath:
     kind: Metric
-    parsed_field: pop2000+pop2008   - pop2000 * pop2008 /pop2000
+    field: pop2000+pop2008   - pop2000 * pop2008 /pop2000
